refactor(mask): opt code
